Texture overhaul; render distance dithering and alternate terrain model
So this is a relatively big update after a long time.
The PBR textures have been reworked to reduce the amount by 50%, by combining the specular map into the albedo, and the heightmaps into the normal. down from 36 textures (9 materials, 4 textures each) to 18 (9x2).

The second change is to the pixel shaders, and the render distance - 6 neighbouring chunks are now rendered  instead of the closest 3; and there is a dithering fade effect towards the render distance (increased to ~5000)

lastly there is a new, alternate terrain algorithm which is currently WIP. trying to adapt the Blender noise algorithms (distortion specifically) into the generation method.